requirements.txt
Flask==3.0.3
numpy==1.26.4
Pillow==10.0.1
tensorflow==2.15.0
mtcnn==0.1.1
paho-mqtt==1.6.1